Move ghost rendering to separate component
